feat: add icons to WorldEdit integration menu
